Move GCRoot back into ClrMD
Feeling better about the code, putting back into the library.
