+++ src/Microsoft.Diagnostics.Runtime.Utilities/Microsoft.Diagnostics.Runtime.Utilities.csproj
@@ -41,4 +41,8 @@
   <ItemGroup>
     <ProjectReference Include="..\Microsoft.Diagnostics.Runtime\Microsoft.Diagnostics.Runtime.csproj" />
   </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="GCRoot\" />
+  </ItemGroup>
 </Project>
