+++ src/Microsoft.Diagnostics.Runtime/src/Common/GCRootPath.cs
