+++ src/Microsoft.Diagnostics.Runtime/src/Implementation/ParallelObjectSet.cs
@@ -4,7 +4,7 @@
 
 using System;
 
-namespace Microsoft.Diagnostics.Runtime
+namespace Microsoft.Diagnostics.Runtime.Implementation
 {
     internal class ParallelObjectSet : ObjectSet
     {
